✨ feat: add permissions to athlete controllers and refactor to clean …

+++ apps/api/src/modules/members/athlete/application/use-cases/delete-athlete.use-case.ts
@@ -1,20 +1,41 @@
-import { Inject, Injectable, NotFoundException } from '@nestjs/common';
+import { ForbiddenException, Inject, Injectable, NotFoundException } from '@nestjs/common';
 import { ATHLETE_WRITE_REPO, AthleteWriteRepository } from '../ports/athlete-write.repository';
+import { UserService } from '../../../auth/user.service';
 
 @Injectable()
 export class DeleteAthleteUseCase {
   constructor(
     @Inject(ATHLETE_WRITE_REPO)
-    private readonly athleteWriteRepository: AthleteWriteRepository
+    private readonly athleteWriteRepository: AthleteWriteRepository,
+    private readonly userService: UserService
   ) {}
 
-  async execute(id: string): Promise<void> {
-    const athlete = await this.athleteWriteRepository.ofId(id);
+  async execute(idAthlete: string, userId: string): Promise<void> {
+
+    //1. Get User
+    const user = await this.userService.getUserById(userId);
+    if (!user) {
+      throw new NotFoundException('User not found');
+    }
+
+    //2. Get Athlete
+    const athlete = await this.athleteWriteRepository.ofId(idAthlete);
 
     if (!athlete) {
       throw new NotFoundException('Athlete not found');
     }
-    
+
+    //3. Check if Athlete has a user  
+    if (!athlete.user) {
+      throw new NotFoundException('Athlete has no user');
+    }
+
+    //4. Check if Athlete belongs to User
+    if (athlete.user.id !== userId) {
+      throw new ForbiddenException('Athlete does not belong to User');
+    }
+
+    //5. Delete Athlete
     await this.athleteWriteRepository.remove(athlete);
   }
 }
